requirements.txt
openai-whisper>=20231117
coqui-tts>=0.16.0
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.11.3
anthropic>=0.8.0
openai>=1.3.0
termcolor>=2.3.0
torch>=2.1.0
numpy>=1.24.0
requests>=2.31.0
pandas>=2.1.1
transformers>=4.35.0
sentence-transformers>=2.2.2